(wip: add offer model)

+++ astrobin_apps_equipment/models/equipment_item_marketplace_listing_line_item.py
@@ -69,7 +69,9 @@ class EquipmentItemMarketplaceListingLineItem(HashedSafeDeleteModel):
         null=True,
     )
 
-    price = models.PositiveIntegerField(
+    price = models.DecimalField(
+        max_digits=10,
+        decimal_places=2,
         null=True,
         blank=True,
     )
